Rename type.cpp/hpp to util.cpp/hpp.
