+++ cpp/src/jit/util.cpp
