+++ cpp/src/jit/util.hpp
