+++ cpp/tests/utilities/column_utilities.cu
@@ -35,8 +35,6 @@
 #include <cudf_test/cudf_gtest.hpp>
 #include <cudf_test/detail/column_utilities.hpp>
 
-#include <jit/type.hpp>
-
 #include <rmm/exec_policy.hpp>
 
 #include <thrust/copy.h>
